BACKEND/requirements.txt
about-time==4.2.1
alive-progress==3.1.5
annotated-types==0.7.0
anyio==4.6.2.post1
assemblyai==0.35.1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
grapheme==0.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.26.4
pandas==2.2.2
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
Wave==0.0.2
websockets==14.1
Werkzeug==3.1.3
